move  get_learning_units_year in assistant_form.py + factories
