+++ assistant/templates/tutoring_learning_unit_year.html
@@ -142,7 +142,7 @@
 
 $(function() {
   $("#search_learning_units_year").autocomplete({
-	source: "/assistants/api/get_learning_units_year/",
+	source: "/assistants/assistant/form/part2/get_learning_units_year/",
 	minLength: 2,
     delay: 100,
     select: function(event, ui) {
