+++ assistant/tests/views/test_assistant_form.py
@@ -23,46 +23,43 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
+import datetime
+import json
+
 from django.test import TestCase, RequestFactory
 from django.core.urlresolvers import reverse
-from django.contrib.auth.models import User
-from base.models.person import Person
-from base.models.academic_year import AcademicYear
-from base.models import academic_year
-from datetime import date
+
+from base.tests.factories.academic_year import AcademicYearFactory
+from base.tests.factories.learning_unit_year import LearningUnitYearFactory
+from base.tests.factories.user import UserFactory
+from base.tests.factories.person import PersonFactory
+
 from assistant.views.assistant_form import form_part4_edit
-from assistant.models.academic_assistant import AcademicAssistant
-from assistant.models.assistant_mandate import AssistantMandate
+from assistant.tests.factories.academic_assistant import AcademicAssistantFactory
+from assistant.tests.factories.assistant_mandate import AssistantMandateFactory
 from assistant.tests.factories.settings import SettingsFactory
+from assistant.models.enums import assistant_mandate_state
 
 
 class AssistantFormViewTestCase(TestCase):
 
     def setUp(self):
         self.factory = RequestFactory()
-        self.user = User.objects.create_user(
-            username='assistant', email='laurent.buset@uclouvain.be', password='assistant'
-        )
-        self.user.save()
-        self.client.login(username=self.user.username, password=self.user.password)
-        self.person = Person.objects.create(user=self.user, first_name='first_name', last_name='last_name')
-        self.person.save()
-        self.academic_assistant = AcademicAssistant.objects.create(person=self.person)
-        self.academic_assistant.save()
-        self.academic_year = AcademicYear.objects.create(year=2016, start_date=date(2016, 9, 1),
-                                                         end_date=date(2017, 8, 31))
-        self.academic_year.save()
-        self.current_academic_year = academic_year.current_academic_year()
-        self.assistant_mandate = AssistantMandate.objects.create(assistant=self.academic_assistant,
-                                                                 academic_year=self.current_academic_year,
-                                                                 entry_date=date(2015, 9, 1),
-                                                                 end_date=date(2017, 9, 1),
-                                                                 fulltime_equivalent=1,
-                                                                 renewal_type='normal',
-                                                                 state='TRTS'
-                                                                 )
-        self.assistant_mandate.save()
         self.settings = SettingsFactory()
+        self.user = UserFactory()
+        self.person = PersonFactory(first_name="James", last_name="Dupont", user=self.user)
+        self.academic_assistant = AcademicAssistantFactory(person=self.person)
+        self.assistant_mandate = AssistantMandateFactory()
+        self.assistant_mandate.assistant = self.academic_assistant
+        self.assistant_mandate.state = assistant_mandate_state.TRTS
+        self.academic_year = AcademicYearFactory(year=datetime.date.today().year)
+        self.learning_unit_year = LearningUnitYearFactory(
+            acronym="LBIR1210",
+            academic_year=self.academic_year)
+        self.learning_unit_year = LearningUnitYearFactory(
+            acronym="LBIR1211",
+            academic_year=self.academic_year)
+        self.client.login(username=self.user.username, password=self.user.password)
 
     def test_assistant_form_part4_edit_view_basic(self):
         request = self.factory.get(reverse('form_part4_edit', kwargs={'mandate_id': self.assistant_mandate.id}))
@@ -70,3 +67,18 @@ class AssistantFormViewTestCase(TestCase):
         with self.assertTemplateUsed('assistant_form_part4.html'):
             response = form_part4_edit(request, self.assistant_mandate.id)
             self.assertEqual(response.status_code, 200)
+
+    def test_get_learning_units_year(self):
+        response = self.client.generic(method='get',
+                                       path='/assistants/assistant/form/part2/get_learning_units_year/?term=LBIR1211',
+                                       HTTP_X_REQUESTED_WITH='XMLHttpRequest')
+
+        data = json.loads(response.content.decode('utf-8'))
+        self.assertEqual(data[0]['value'], 'LBIR1211')
+        self.assertEqual(len(data), 1)
+
+        response = self.client.generic(method='get',
+                                       path='/assistants/assistant/form/part2/get_learning_units_year/?term=LBIR12',
+                                       HTTP_X_REQUESTED_WITH='XMLHttpRequest')
+        data = json.loads(response.content.decode('utf-8'))
+        self.assertEqual(len(data), 2)
