+++ assistant/views/assistant_form.py
@@ -25,15 +25,17 @@
 ##############################################################################
 from django.contrib.auth.decorators import login_required, user_passes_test
 from django.core.urlresolvers import reverse
+from django.http import JsonResponse
 from django.http.response import HttpResponseRedirect
 from django.shortcuts import render
 from django.core.exceptions import ObjectDoesNotExist
 from django.utils.translation import ugettext as _
+from base.models import person_address, person, learning_unit_year
+from base.models.learning_unit_year import search
 from assistant.models import *
 from assistant.utils.send_email import send_message
 from assistant.forms import *
 from assistant.models.enums import document_type
-from base.models import person_address, person, learning_unit_year
 from assistant.models.enums import assistant_mandate_state, assistant_phd_inscription
 
 
@@ -49,6 +51,23 @@ def user_is_assistant_and_procedure_is_open(user):
             return False
     except ObjectDoesNotExist:
         return False
+
+
+#@user_passes_test(user_is_assistant_and_procedure_is_open, login_url='access_denied')
+def get_learning_units_year(request):
+    if request.is_ajax() and 'term' in request.GET:
+        q = request.GET.get('term')
+        learning_units_year = search(acronym=q)[:50]
+        response_data = []
+        for learning_unit_year in learning_units_year:
+            response_data.append({'value': learning_unit_year.acronym,
+                                  'title': learning_unit_year.title,
+                                  'academic_year': str(learning_unit_year.academic_year),
+                                  'id': learning_unit_year.id
+                                  })
+    else:
+        response_data = []
+    return JsonResponse(response_data, safe=False)
 
 
 @user_passes_test(user_is_assistant_and_procedure_is_open, login_url='access_denied')
@@ -98,7 +117,7 @@ def tutoring_learning_unit_edit(request, tutoring_learning_unit_id=None):
     mandate_id = edited_tutoring_learning_unit_year.mandate_id
     mandate = assistant_mandate.find_mandate_by_id(mandate_id)
     search_learning_units_year = edited_tutoring_learning_unit_year.learning_unit_year.acronym + \
-                                 ' (' + str(edited_tutoring_learning_unit_year.learning_unit_year.academic_year) + ')'
+        ' (' + str(edited_tutoring_learning_unit_year.learning_unit_year.academic_year) + ')'
 
     form = TutoringLearningUnitForm(
         initial={
@@ -127,14 +146,14 @@ def tutoring_learning_unit_edit(request, tutoring_learning_unit_id=None):
 def tutoring_learning_unit_save(request):
     if request.method == 'POST':
         tutoring_learning_unit_year_id = request.POST.get('tutoring_learning_unit_year_id')
+        mandate_id = request.POST.get('mandate_id')
+        learning_unit_year_id = request.POST.get('learning_unit_year_id')
         if tutoring_learning_unit_year_id:
             current_tutoring_learning_unit = mdl.tutoring_learning_unit_year.find_by_id(tutoring_learning_unit_year_id)
         else:
             current_tutoring_learning_unit = None
         form = TutoringLearningUnitForm(data=request.POST, instance=current_tutoring_learning_unit)
-        if form .is_valid():
-            mandate_id = request.POST.get('mandate_id')
-            learning_unit_year_id = request.POST.get('learning_unit_year_id')
+        if form.is_valid():
             this_mandate = assistant_mandate.find_mandate_by_id(mandate_id)
             current_tutoring_learning_unit = form.save(commit=False)
             if not learning_unit_year_id:
@@ -146,7 +165,7 @@ def tutoring_learning_unit_save(request):
                 current_tutoring_learning_unit.mandate = this_mandate
                 current_tutoring_learning_unit.save()
                 return HttpResponseRedirect(reverse('mandate_learning_units', kwargs={'mandate_id': mandate_id}))
-    return render(request, "tutoring_learning_unit_year.html", {'form': form, 'mandate_id': mandate_id})
+        return render(request, "tutoring_learning_unit_year.html", {'form': form, 'mandate_id': mandate_id})
 
 
 @login_required
@@ -217,11 +236,9 @@ def form_part3_save(request, mandate_id):
 @user_passes_test(user_is_assistant_and_procedure_is_open, login_url='access_denied')
 def form_part4_edit(request, mandate_id):
     mandate = assistant_mandate.find_mandate_by_id(mandate_id)
-    person = request.user.person
+    current_person = request.user.person
     assistant = mandate.assistant
     files = assistant_document_file.find_by_assistant_mandate_and_description(mandate, document_type.RESEARCH_DOCUMENT)
-    if person != assistant.person or mandate.state != assistant_mandate_state.TRTS:
-        return HttpResponseRedirect(reverse('assistant_mandates'))
     form = AssistantFormPart4(initial={'internships': mandate.internships,
                                        'conferences': mandate.conferences,
                                        'publications': mandate.publications,
